pyproject.toml
[project]
name = "data-masker"
version = "0.1.0"
description = "Data Masker – Smart Data Redactor"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
	"pandas",
	"click",
	"pyyaml",
	"openpyxl",
]

[project.optional-dependencies]
dev = [
	"pytest",
	"ruff",
	"mypy",
	"pre-commit",
]

[project.scripts]
data-masker = "data_masker.cli:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["data_masker*"]
exclude = ["tests*", "sample_data*"]

[tool.ruff]
line-length = 100
select = ["E","F","W","I","UP","B","SIM","PL" ]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["data_masker"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
